Separated the managers on it's own files
