+++ app/repositories/managers/ingredient_manager.py
@@ -1,0 +1,13 @@
+from typing import Sequence
+
+from .base_manager import BaseManager
+from ..models import Ingredient
+from ..serializers import IngredientSerializer
+
+class IngredientManager(BaseManager):
+    model = Ingredient
+    serializer = IngredientSerializer
+
+    @classmethod
+    def get_by_id_list(cls, ids: Sequence):
+        return cls.session.query(cls.model).filter(cls.model._id.in_(set(ids))).all() or []
