+++ app/repositories/managers/order_manager.py
@@ -1,0 +1,25 @@
+from typing import List
+from .base_manager import BaseManager
+from ..models import Ingredient, Beverage, Order, IngredientOrderDetail, BeverageOrderDetail
+from ..serializers import OrderSerializer
+
+class OrderManager(BaseManager):
+    model = Order
+    serializer = OrderSerializer
+
+    @classmethod
+    def create(cls, order_data: dict, ingredients: List[Ingredient], beverages: List[Beverage]):
+        new_order = cls.model(**order_data)
+        cls.session.add(new_order)
+        cls.session.flush()
+        cls.session.refresh(new_order)
+        cls.session.add_all((IngredientOrderDetail(order_id=new_order._id, ingredient_id=ingredient._id, ingredient_price=ingredient.price)
+                             for ingredient in ingredients))
+        cls.session.add_all((BeverageOrderDetail(order_id=new_order._id, beverage_id=beverage._id, beverage_price=beverage.price)
+                             for beverage in beverages))
+        cls.session.commit()
+        return cls.serializer().dump(new_order)
+
+    @classmethod
+    def update(cls):
+        raise NotImplementedError(f'Method not suported for {cls.__name__}')
