+++ app/repositories/managers/size_manager.py
@@ -1,0 +1,7 @@
+from .base_manager import BaseManager
+from ..models import Size
+from ..serializers import SizeSerializer
+
+class SizeManager(BaseManager):
+    model = Size
+    serializer = SizeSerializer
